# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"body": {}
				},
				"required": [
					"status",
					"body"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_any-Array_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"body": {
						"items": {},
						"type": "array"
					}
				},
				"required": [
					"status",
					"body"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-content-api",
		"version": "1.0.0",
		"license": {
			"name": "UNLICENSED"
		},
		"contact": {
			"name": "Tiago Redaelli ",
			"email": "tiago.redaelli@if.se"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/content/{id}": {
			"get": {
				"operationId": "GetContent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "editmode",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/content/{id}/children": {
			"get": {
				"operationId": "GetContentChildren",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any-Array_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "editmode",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/content/pages": {
			"get": {
				"operationId": "AllPages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "editmode",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/content/pages/tree": {
			"get": {
				"operationId": "AllPagesAsTree",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "editmode",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/content/pages/{id}": {
			"get": {
				"operationId": "GetPage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "editmode",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/content/pages/{id}/children": {
			"get": {
				"operationId": "GetPageChildren",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "editmode",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/content/sites": {
			"get": {
				"operationId": "GetWebsites",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}